requirements.txt
# Data manipulation
pandas
numpy

# Machine learning and Baselines
scikit-learn
xgboost
statsmodels

# Deep Learning
pytorch

# Visualization
matplotlib
seaborn